fix: Match current production schema

+++ app/Modules/Messages/Database/Migrations/2019_12_11_000000_create_messagequeue_tables.php
@@ -28,7 +28,7 @@ class CreateMessagequeueTables extends Migration
 				$table->dateTime('datetimestarted')->nullable();
 				$table->dateTime('datetimecompleted')->nullable();
 				$table->integer('pid')->unsigned()->default(0);
-				$table->integer('returnstatus')->unsigned()->default(0);
+				$table->smallInteger('returnstatus')->unsigned()->default(0);
 				$table->string('output', 150);
 				$table->index('datetimecompleted');
 				$table->index(['targetobjectid', 'userid'], 'targetobject');
